package/golang: new package
This patch adds golang, the Go compiler-and-interpreter.

We're adding a host package to get the golang cross-compiler, and a
target package to get the golang interpreter on the target.

It is to be noted that golang's buildsystem is going at great length to
not do anything like any other buildsystem, and is full of little
tricks.

First-off, even though it generates code for the target, the host-variant
does not need the toolchain; it has its own built-in code generator. Yet,
building the target variant, the toolchain is indeed required. Weird...

Second, it misdetects cross-compilation: it considers that if the target
CPU and the build CPU are the same, then this is not cross-compilation.
Of course, this is completely stupid and wrong. So, we need to patch it
to override its insane heuristic.

Third, even so, when the target CPU is the same as the build CPU, the
resulting binaries will be dynamically linked, otherwise, they are
statically linked. We can't even force it one way or the other...
Instead, we mark it as requiring dynamic libraries, and let the golang
ill buildsystem decide whether to do static or dynamic link; we can do
so because any sane toolchain will always have the static variants of
the libc and libpthread, which is all that golang requires anyway.

Signed-off-by: Christian Stewart <christian@paral.in>
[yann.morin.1998@free.fr:
 - add architecture-specific options
 - needs dynamic linking
 - fix licensing info
 - fix tab-space indentation in patch 0003
 - fix offset in patch 0001
 - misc cleanups
]
Signed-off-by: Yann E. MORIN <yann.morin.1998@free.fr>

---
Changes v6 -> v7;
  - added variable for cross-compiling env vars & command line
  - added additional comments and documentation differentiating runtime
  - fix case of packge name comment
Changes v4 -> v6;
  - origin of hash  (Thomas)
  - better indent commands  (Thomas)
  - don't "deinstall" prior to installing  (Thomas)
  - fix offset in patch 0001

Changes v3 -> v4;
  - adopted by Yann, worked on in concert with Christian
  - lots of refactoring and cleanups